tests: share one corpus-runnability gate, fixing the core-only legs
CI was red on ubuntu core-only (py3.11, py3.14) and the free-threaded leg:
6 tests in the review guardrails died with ModuleNotFoundError: py7zr,
while every [all] leg passed.

Cause: the guardrails' local _archive() helper gated on reader availability
only. That is exactly wrong for 7z -- the reader is native, so
format_availability reports the format fine, but the corpus *builds* 7z with
py7zr, which core-only does not install. Two more copies of the same
hand-rolled gate had been pasted into the encrypted-header and data-encrypted
tests; the second survived only because it happens to pick zip first.

test_corpus_sweep already had the correct gate, consulting READER_PACKAGES,
BUILDER_PACKAGES and BUILDER_BINARIES together -- and test_mutation_fuzz and
test_short_read_sources were already reaching across module boundaries to
import it privately. So the fix is to promote it rather than to copy it a
fourth time: skip_unless_runnable now lives in sample_archives, next to the
three tables it reads, and all four call sites use it. pytest is imported
lazily there because probe_matrix.py imports the module outside pytest.

Also corrects this file's own claim that the three-config rule was "a
formality here". It is not, and this is the demonstration: no *finding*
depends on an optional library, but the tests that pin the findings do.
WORKPLAN.md now carries the exact commands to reproduce the reduced legs and
a note to use the shared gate instead of format_availability alone.

Verified locally in both configurations:
- core-only (uv sync --no-dev): 1753 passed, 451 skipped, 13 xfailed
- all (uv sync --group dev --extra all): 2193 passed, 65 skipped, 16 xfailed
ruff, pyrefly and ty clean.
